submodules/requirements.txt
mysql-connector-python
simple-term-menu